[Minor] clean up multimodal processor and tokenizer manager (#7624)
